requirements.txt
Django==3.0.7
django-extensions==3.0.9
django-jinja==2.6.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2client==3.0.0
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-jwt==2.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
